Rename Mancala to lowecase file name
